Renamed searchDoiQuery file
