+++ src/data/queries/searchDoiQuery.ts
